componentizando pagina principal
