+++ components/Events/index.ts
@@ -1,0 +1,1 @@
+export {default as Events} from './Events'
